api/pyproject.toml
[tool.poetry]
name = "spotifyle-api"
version = "0.1.0"
description = "Backend API for the Spotifyle web based game."
authors = ["dgnsrekt <dgnsrekt@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
django-ninja = "^0.16.1"
django-environ = "^0.8.1"
django-cors-headers = "^3.10.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
celery = {extras = ["redis"], version = "^5.2.2"}
requests-oauthlib = "^1.3.0"
Faker = "^10.0.0"
faker_music = "^0.4"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.4"
pytest-watch = "^4.2.0"
isort = "^5.10.1"
pytest-clarity = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
